+++ .next/static/chunks/webpack.js
@@ -161,7 +161,7 @@
 /******/ 		// This function allow to reference async chunks
 /******/ 		__webpack_require__.u = (chunkId) => {
 /******/ 			// return url for filenames based on template
-/******/ 			return undefined;
+/******/ 			return "static/chunks/" + chunkId + ".js";
 /******/ 		};
 /******/ 	})();
 /******/ 	
@@ -190,7 +190,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("75014dcd22c67eec")
+/******/ 		__webpack_require__.h = () => ("fb2554b46a591931")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */